+++ cmem_plugin_reason/plugin_reason.py
@@ -2,6 +2,7 @@
 
 from collections.abc import Sequence
 from datetime import UTC, datetime
+from os import environ
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from time import time
@@ -32,6 +33,8 @@ from cmem_plugin_reason.utils import (
     send_result,
     validate_profiles,
 )
+
+environ["SSL_VERIFY"] = "false"
 
 
 @Plugin(