+++ cmem_plugin_reason/plugin_validate.py
@@ -1,6 +1,7 @@
 """Ontology consistency validation workflow plugin module"""
 
 from datetime import UTC, datetime
+from os import environ
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from time import time
@@ -37,6 +38,8 @@ from cmem_plugin_reason.utils import (
     send_result,
     validate_profiles,
 )
+
+environ["SSL_VERIFY"] = "false"
 
 
 @Plugin(